[REFACTOR] 알림 관련 기능 cqrs 패턴으로 수정 및 로직 리팩토링

+++ src/main/java/spring/hi_hello_spring/notify/command/application/service/NotifyServiceImpl.java
@@ -1,4 +1,4 @@
-package spring.hi_hello_spring.notify.service;
+package spring.hi_hello_spring.notify.command.application.service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -13,15 +13,16 @@ import spring.hi_hello_spring.common.exception.ErrorCodeType;
 import spring.hi_hello_spring.common.util.CustomUserUtils;
 import spring.hi_hello_spring.employee.command.domain.aggregate.entity.Employee;
 import spring.hi_hello_spring.employee.command.domain.repository.EmployeeRepository;
-import spring.hi_hello_spring.notify.dto.NotifyDTO;
-import spring.hi_hello_spring.notify.entity.NotiType;
-import spring.hi_hello_spring.notify.entity.Notify;
-import spring.hi_hello_spring.notify.repository.EmitterRepositoryImpl;
-import spring.hi_hello_spring.notify.repository.NotifyRepository;
+import spring.hi_hello_spring.notify.command.domain.aggregate.entity.NotiType;
+import spring.hi_hello_spring.notify.command.domain.aggregate.entity.Notify;
+import spring.hi_hello_spring.notify.command.domain.repository.EmitterRepositoryImpl;
+import spring.hi_hello_spring.notify.command.infrastructure.repository.JpaNotifyRepository;
+import spring.hi_hello_spring.notify.command.application.dto.NotifyDTO;
 
 import java.io.IOException;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -31,7 +32,7 @@ public class NotifyServiceImpl implements NotifyService {
     // SSE 연결 지속 시간 설정 (1시간)
     private static final Long DEFAULT_TIMEOUT = 60L * 1000 * 60;
 
-    private final NotifyRepository notifyRepository;
+    private final JpaNotifyRepository notifyRepository;
     private final EmitterRepositoryImpl emitterRepository;
     private final EmployeeRepository employeeRepository;
 
@@ -115,6 +116,7 @@ public class NotifyServiceImpl implements NotifyService {
 
     public void sendToReceivers(Employee sender, List<Employee> receivers, NotiType notiType, String url) {
 
+        // 알림 생성 및 저장
         receivers.forEach(receiver -> {
             String senderSeq = String.valueOf(sender.getEmployeeSeq());
             Notify notification = notifyRepository.save(
@@ -135,6 +137,7 @@ public class NotifyServiceImpl implements NotifyService {
                 objectMapper.registerModule(new JavaTimeModule());
                 String jsonData = null;
                 try {
+                    // 알림 응답 DTO로 매핑하도록 수정해야함
                     jsonData = objectMapper.writeValueAsString(NotifyDTO.createResponse(notification, employeeSeq));
                 } catch (JsonProcessingException e) {
                     throw new RuntimeException(e);
@@ -155,36 +158,28 @@ public class NotifyServiceImpl implements NotifyService {
                 .build();
     }
 
-    // 알림 전체 조회
-    @Transactional
-    public List<NotifyDTO> notiAll(Long employeeSeq){
-
-        return notifyRepository.findAllByEmployeeSeq(employeeSeq)
-                .stream()
-                .map(NotifyDTO::new)
-                .toList();
-    }
-
     // 알림 단일 읽음
     @Transactional
     public void readNotify(Long employeeSeq, Long notiSeq){
 
-//        Notify notify = checkNotifyAndEmployeeSeq(employeeSeq, notiSeq);
-        Notify notify = notifyRepository.findById(notiSeq).orElse(null);
-
-        // 읽음 처리
-        notify.notiRead();
+        Notify notify = notifyRepository.findById(notiSeq)
+                .orElseThrow(() -> new CustomException(ErrorCodeType.DATA_NOT_FOUND));
+        if (Objects.equals(notify.getEmployeeSeq(), employeeSeq)) {
+            // 읽음 처리
+            notify.notiRead();
+        }
     }
 
     // 알림 단일 삭제
     @Transactional
     public void deleteNotify(Long employeeSeq, Long notiSeq){
 
-//        Notify notify = checkNotifyAndEmployeeSeq(employeeSeq, notiSeq);
-        Notify notify = notifyRepository.findById(notiSeq).orElse(null);
-
-        // 삭제 처리
-        notifyRepository.delete(notify);
+        Notify notify = notifyRepository.findById(notiSeq)
+                .orElseThrow(() -> new CustomException(ErrorCodeType.DATA_NOT_FOUND));
+        if (Objects.equals(notify.getEmployeeSeq(), employeeSeq)) {
+            // 삭제 처리
+            notifyRepository.delete(notify);
+        }
     }
 
     // 알림 전체 읽음
